refactor: 팀 선정 모달 위치 TableOverview page 폴더 내부로 이동
